requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
branca==0.7.2
celery==5.3.1
certifi==2024.6.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.0.0
Django==4.2.3
django-celery-beat==2.6.0
django-cors-headers==4.1.0
django-filter==23.2
django-redis==5.3.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
Faker==18.11.1
folium==0.14.0
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
zope.interface==6.4.post2
